Create Inventory Manager
Created inventory manager to handle list of inventory objects. Fleshed out scripts to add functions and variables to handle everything for inventory system. Added to text prompt SO and Game Manager to allow for new functionality. Used new functionaliuty to allow hat and key to be added to inventory through text prompts.